refactor(boot): extract boot sequence use case to core
