+++ core/src/main/java/com/marmitt/core/dto/boot/BootExecutionSummary.java
@@ -1,0 +1,12 @@
+package com.marmitt.core.dto.boot;
+
+import com.marmitt.core.application.usecase.runner.RunnerBootRecoveryUseCase;
+
+import java.util.List;
+
+public record BootExecutionSummary(
+        int portfoliosCount,
+        int eligibleRunnersCount,
+        List<RunnerBootRecoveryUseCase.RecoverySummary> runnerSummaries
+) {
+}
